Made component library analytics site-wide
